Don't try to update a node in power transition
If a node is registered while it is turned on, it may take some time
for Ironic to assert the power state, and while it is doing so it
won't allow us to update the node.  This can be a problem for
the "configure boot" command because it would typically be run right
after registration.

To avoid this, make sure the node's power state has been updated
before trying to do the update.  Note that I did not just leave this
to the Ironic client retry logic because it results in a large
number of warnings logged without any explanation why.  Explicitly
checking the state allows us to log a more descriptive message.

Change-Id: I8c349752aed98645615b06957cbea90b8d63039f