Pass extra heat templates like heatclient
Currently, to pass extra heat templates to the deployment,
all of the templates need to be placed in a subdirectory
of /etc/tripleo/extra_config.d/, and they must either be
named *environment*.yaml or *registry*.yaml. This was done
specifically so we could just drop the rhel-registration
templates directory from tripleo-heat-templates in our
extra config drop directory, and it would just work.

However, after trying to explain how this interface works
to folks adding other extra configuration, it seems like
this was not the best option. Instead of the drop directory
approach, we just pass all extra configuration with
individual -e options, exactly how it is done via heatclient.

Change-Id: Ic0f6a69dd7f71d2777bbf6f07f1a21e076064553