Increase the sleep time while trying to get neutron l3 agents
After deploy and before postconfig increase the sleep time
before trying to initialize neutron (and poll for enough
l3 agents). Also prints the warnings to stdout so there is more
immediate/useful info if a deploy is taking a while.

This is an update to
https://review.gerrithub.io/#/c/238097/
and more details can be found on the bug report at
https://bugzilla.redhat.com/show_bug.cgi?id=1236578

Change-Id: I8d59b1c01f9a5d3174acfd6f2d2f54e7cf62904d
Related-Bug: https://bugzilla.redhat.com/show_bug.cgi?id=1236578