Updating overcloud should reuse heat parameters
When doing an overcloud deploy, the unified cli already detects an
existing overcloud and performs a heat stack update instead of
another create. This change makes the stack update get performed
with reuse of the existing parameters already passed to heat when
the stack was created. Without this the parameters for the keystone
ssl setup were not getting put into the parameters for the update,
causing it to fail.

See also: https://bugzilla.redhat.com/show_bug.cgi?id=1240123

Change-Id: Iba051d5a611ee69f9e6ce4f1a3468d377d542408