Fix the deploy tests
The ManillaPassword parameter was added without the tests being updated.
This caused a number of failures. This change also sorts the parameters
to make them easier to visiaully inspect and adds an assert to check the
symetric difference between the parmater keys - making it easier to see
what the extra parameter is when there is one.

Change-Id: I75542ca1bb457ef55405235cbc481b6811a112a8