image build: prepare for different generation methods
Isolate the bits creating the command line invocations to
disk-image-create/ramdisk-image-create into a small class;
introduce a new parameter to decide which method to use for generating
the images, with the only value of 'dib', representing the current
diskimage-builder method.

Force the use of dib in TestOvercloudImageBuild, and mock the methods of
DibImageBuilder.

Change-Id: I9287cd9375fbc496f3d2bb570a61cd320a2e451b