Build IPA in place of discoverd ramdisk
The ironic-discoverd packaging included python code to be
run in a ramdisk for gathering hardware details.

However, upstream is moving to ironic-python-agent, so the
python ramdisk code is no longer packaged with the move to
ironic-inspector.

This patch resolves this by switching to the IPA ramdisk
for inspection. This will also require a patch to
instack-undercloud.

Depends-On: I476f9ec9ec4206ece0261eaaf2b4182c6bcbd802
Change-Id: I73d39aca2410f0c2e5ead8e1b59d57af053bade7
Partial-Bug: 1498068