Don't send default environment file on stack update
With the improvements for bug #1224828 in Liberty, Heat now remembers
the stack's environment so there is no need to pass all environment
files on every update of overcloud.

Since user can now omit all the '-e' arguments on update operations
we also should not send the "basic" resource registry environment
file either because if only this file is passed on update operation
(without all the extra env files passed on create), it may cause that
resource_registry for some resources will be changed which causes
that existing resources to be replaced with new ones.

Change-Id: I2e2da03e6a36f3bc46bb608350e11f5431e0fcff
Closes-Bug: #1498607