Fix file checksum function in for python 3.4
The file_checksum function in tripleoclient/utils.py resulted in an
infinite loop in py34 due to the fact that that function was expecting
an empty character to finish the iteration. This is no longer the case
in py34. So instead, we now iterate the file until it can no longer read
anything. Even though it's more verbose; this works for both python 2.7
and 3.4.

Change-Id: I4849182992cac1209431a2147a83c9167d6a8af2
Closes-Bug: #1500592