Install device-mapper-multipath on overcloud-full
device-mapper-multipath needs to be installed for managing of multipath
devices for scenarios where fail over or load balancing of block devices
is desirable.

This change installs the package by default in the overcloud-full image
for connvenience. The multipathd daemon is not started or enabled nor
any multipath configuration exposed. Those exercises would be left to be
done via the ExtraConfig resources in tripleo-heat-templates.

Change-Id: I5b575fe7728fd58c1fd448b1f4355ee4df7bffb8