Handle images not being found
If the expected deployment kernel and ramdisk images are not found, the
deployment will fail. Technically, this is a validation error, so should
be appropriately handled via the --validation-errors-fatal cli arg.

There are reasons why such a validation should not be fatal, for
instance if the deployment images had different names than the hardcoded
ones here, etc.

Change-Id: I0bffd513b3eb79813df3971fb990950ee2bde65d