Set --min-tmpfs for overcloud-full build
I consistently have problems building images on an 8 GB VM with
the undercloud already installed (which is the documented flow).
At 4 or 12 GB it works fine, but 8 seems to fall in a sweet spot
where the tmpfs usage of dib causes OOM problems.  4 GB works
because dib doesn't use tmpfs by default with that little memory,
and 12 GB seems to be sufficient to avoid running out of memory
even with tmpfs.

This changes passes the --min-tmpfs 5 parameter to the image build
for the overcloud-full image (the largest image by far) so it will
not attempt to build in tmpfs if there is less than 12 GB of RAM.
5 instead of 6 because due to rounding 6 would result in tmpfs not
being used on 12 GB VMs either (similar to how it is not used on a
4 GB VM even though the default is 2).

Note that tmpfs by default uses half of total memory, which is why
all of the min-tmpfs numbers are half of the expected memory on the
system.

Change-Id: I6254d04216936551e68ac02999f2bffc11b19f65
(cherry picked from commit 608b06976281aee5f079ddaa2207691f3160c087)