Don't download fedora-user as part of --all build
fedora-user is an optional image that is only used for testing the
deployed overcloud.  We shouldn't force everyone to download it as
part of the --all build when in many cases it won't be needed.  The
option to download it is left in the client though so it can still
be easily retrieved when necessary.

Change-Id: I3ea6ad911a6ee411f19379060da127ef0a6e75e3