Log all resource events during overcloud deploy
This change switches the overcloud deploy stack polling with event
list polling. This uses the same technique as the recently added
heat stack-create --poll mode: Ib7d35b66521f0ccca8544fd18fb70e04eaf98e5a

There is a (CREATE|UPDATE)_(COMPLETE|FAILED) event generated when the stack
operation is complete, so this is used to decide when to stop polling.

Also, every time there is an event it is printed to the screen in log
style. This would give the user an indication that something is
happening, which is better than staring at nothing for minutes (hours).

On stack update, the oldest event from the previous operation is discovered and
that is used for the marker so that only new events are logged.

Change-Id: I8951b0b924db769961aa67e654fcbd774d3e7787
(cherry picked from commit a6ff771ab773e5214f46c1a86d8d829d74903e16)