Support new ironic "enroll" state
This state was introduced in Liberty as a new state for freshly
enrolled node. Transition from it is done by the same verb "manage",
but now involves validation of power credentials. Nodes are not set to
"available" if they fail introspection or power credentials.

The code changes required some additional changes to test and utility
code, as follows:

 - utils.wait_for_provision_state now raises appropriate exceptions on
   error rather than returning False
 - there's now a utils.nodes_in_states function to list baremetal nodes
   in a given set of states.
 - logic is added to baremetal/fakes.py to track states of fake nodes;
   this is much simpler and less brittle than precisely mocking the
   results of API calls in exactly the order the library code makes
   them.
 - bulk introspection tests consistently mock at the client layer,
   rather than at a variety of layers.
 - timeout and power-credential error during bulk introspection are
   tested.

Change-Id: I4da61491f60f7ebd42ca1f8fe45c3d4df6e49887
Co-Authored-By: Miles Gould <mgould@redhat.com>