Fix chicken and eggs problem with tripleo-heat-templates
There is a current effort on tripleo-heat-templates to manage keystone
resources directly via puppet. The effort is separate in two patchset
one to replace the current keystone.initialize() and the other other to
replace setup-endpoints().

This review is necessary in order to make CI pass the upstream jobs.
The whole keystone functions call will be removed then.

Change-Id: Ia4b3244f114dcff746ab89d355ad4933f8fdbddf