Don't log exceptions redundantly
The pattern

self.log.error('Some message')
self.log.exception(e)

is redundant and misuses the logging.exception function.  The
argument to exception() is supposed to be a string describing the
error that occurred, not the exception itself.  It already pulls
the exception details from the exception handler context and
includes them in the output.

This changes the instances of the above pattern to simply

self.log.exception('Some message')

which results in the correct output of something like:

ERROR:root:Some message
Traceback (most recent call last):
  File "./test.py", line 23, in <module>
    raise Exception('foo')
Exception: foo

Change-Id: Ia55ec34256ceb6f095c2fbfe2d085567f6382509