Allow adding new overcloud passwords
The current implementation of the password generation function is
all or nothing.  If the file exists, then it uses that file as the
list of passwords, otherwise it generates an entirely new file.

This means that if we add a new service that requires a password to
be generated and a user upgrades with a password file generated by
an older version, we will not generate a new password for the new
service and pandemonium will result.

To address the problem, read in any existing passwords from the file,
but loop through the names to make sure all of the expected passwords
are set.  If any are not, generate them.  Behavior in the case of
no password file remains unchanged.

Change-Id: Id9c5fe363a2aef55800bee4db12cbf3e2d278e45