Don't raise bare Exceptions
It's generally recommended that base Exceptions not be raised,
because this makes it impossible to handle exceptions in a more
granular way.  except Exception: will catch all exceptions, not
just the one you might care about.

This replaces most instances of this pattern in tripleoclient, with
the exception of the one in overcloud_image.py because I have a
separate change open that already fixes it.

Change-Id: I6fce306c4ffc57b4c52389be1feb583f5a400a64