Include registry on update when other envs are specified
On an openstack overcloud deploy which results in a stack-update,
implicitly include overcloud-resource-registry-puppet.yaml if the user
specifies any other custom environments. Do not include
overcloud-resource-registry-puppet.yaml if the user specifies no extra
environments.

This solution would be an improvement over the current situation, but it
isn't perfect.

It does the right thing when:
- the user does an openstack overcloud with no environments specified
- the user does an openstack overcloud with all custom environments
  specified

It will sometimes do the wrong thing when:
- the user specifies a partial list of custom environments (like -e
  do_my_custom_oneoff_thing.yaml)

Change-Id: I60fbf2182429f6bb83e000fc4d45269d7f83171b
Related-Bug: #1498607
Closes-Bug: #1504363