Add overcloud deploy --environment-directory
This adds a new --environment-directory option that can be
used to load multiple environment directories (in order specified).
Each directory that is specified is searched for .yaml files which
are then added to the heat stack create/update commands in
front of all the others.

By default the $HOME/.tripleo/environments directory is also
loaded. This location provides a nice way to auto-generate
heat environments in this directory which can then be
automatically included by tripleoclient deployment commands.

Implements: blueprint puppet-modules-deployment-via-swift

Change-Id: Ib2a2db1533e41438a09d393ca6b68c769719d40e