Use renamed overcloud.yaml template
The overcloud-without-mergepy was renamed to overcloud. So this commit
takes it into use. On the other hand, backward compatibility is kept,
the script will first check for the old overcloud-without-mergepy.yaml
and if that's not available, it will then use overcloud.yaml, which is
the what it will use be once we remove overcloud-without-mergepy from
the t-h-t repo.

Change-Id: I94f3048868edc06249f487380c86db7cd400ce4c
Depends-On: I14f5ff78b083b34590d30357df94c42ff6a0c2c0