Fail the introspection command if introspection has failed
Currently it only prints an error message, but exists with success.
This makes it impossible to use this command in any kind of automated
scripts or test it in our gate. This patch makes it raise an exception.

Change-Id: I150c87252a48a8062aa7ef04c7a52433dd5ee37d
(cherry picked from commit ad470c343e4cafd975c8c313d04e762d47c32309)