Enable TLS endpoints through os-cloud-config
Currently the way to configure keystone endpoints is through the
os-cloud-config scripts. This change enables the necessary parameters
in order to get os-cloud-config to configure the keystone endpoints
with TLS enabled. Even though setting up the endpoints in such a way
will be deprecated in favor of doing it through puppet, This patch
enabled a way of enabling TLS that can be backported.

Change-Id: Ie2428093b270ab8bc19fcb2130bb16a41ca0ce09