Allow node import via yaml not only csv/json
JSON, and to a lesser extent CSV, are not friendly formats for human
users, so also support yaml.

Also make the --json/--csv options deprecated and optional - we can
just detect the filetype via the suffix instead of introducting a
new option enabling yaml support.

Note there are future improvements we can make around the test here,
e.g reducing duplication via testscenarios and improving the handling
of temporary files, but I've not considered that in this patch and
have stuck with the current conventions.

Change-Id: I8c5518d71c2db0d56af443c97212f5a63a36a545