Allow to use the old deploy ramdisk
While defaulting to IPA is a correct choice, for the transition period
let's allow people to use the old ramdisk. Adds new option --old-deploy-image
to the upload command and move the deprecation message to it.

This change does not correspond to a Mitaka change, as in Mitaka the old
ramdisk should not be supported.

Change-Id: I311843ed95d23dfdc83d184a09c52f0f5e1b31df
Closes-Bug: #1530883