Simplify checking for stack complete
This is a small simplification over Change-Id:
Id325e793a510e751ec4736883f6f97cb80a5623f
that should also not print out duplicated event lists.

This fixes the issue where the last event is seen, but the overcloud
stack is not yet CREATE_COMPLETE, which caused an infinte loop in the
client.

Change-Id: Idb5b02435667d6ed504e60a4ef6779ee4191c726