Use values from endpoint map for service endpoints
Instead of using values hardcoded in os-cloud-config. This commit
uses the values that come from the endpoint map that's used
in the t-h-t. This in turns, allows us to be able to change the
port and even the host used for the public and internal endpoints;
this is still not the case for the admin endpoints, as it is based
on the internal endpoint of the service. The aforementioned
limitation comes from os-cloud-config.

conflicts:
    tripleoclient/tests/test_utils.py
    tripleoclient/tests/v1/overcloud_deploy/fakes.py
    tripleoclient/utils.py
    tripleoclient/v1/overcloud_deploy.py
    tripleoclient/tests/v1/overcloud_deploy/test_overcloud_deploy.py

Change-Id: I0cf9ada1110f6b1452480039c136aee1fc3ccc82
Closes-Bug: #1569851
(cherry picked from commit d67ccaf7baf7beb7ea3559c84ae0c645d0452dd8)