Refresh node power state at start of loop
When configuring boot on the baremetal nodes, we first do a list
and then iterate over the results.  This can be a minor hassle if
some of the nodes start out in a transitional power state because
the initial list results don't get updated before checking each
subsequent node.  We end up waiting the sleep time unnecessarily
until the node status is refreshed for every node that was in
transition, even though most of them have probably finished by the
time the first one or two do.

Also removes a later call to node.get since it is now
redundant.

Change-Id: I88d4c57167729cbfb24a9c2ca554738a7fbac3ab