Fix ironic-inspector client usage
Right now we are using the ironic-inspector-client functions directly with
providing the auth token. This is deprecated in favor of providing the session.
However, inside the OSC plugin the introspection client is already available,
so just use it directly. This patch also stops redefining --inspector-url which
is already defined by ironic-inspector-client.

Change-Id: I9538e32967103f71c2af94077eb1994462a88d18