Allow 'openstack baremetal configure boot' to guess the root device
This change simplifies using Ironic root device hints for people who only need
to change the default strategy of selecting the default device.
E.g. we have a use case for selecting the first device instead of the smallest,
so that the root device does not change after upgrade Kilo -> Liberty.

Note that this feature does not replace per-node device hints, rather
complement them with a more global and easy to use setting.

This change introduces 3 new arguments to the command:
--root-device states how to find a root devices for a node.
  If this argument is provided, the client will try to detect the root device
  based on the stored introspection data. Possible values:
  * smallest (the same thing as IPA does by default - the smallest device)
  * largest (the opposite thing - pick the largest device)
  * otherwise it's treated as a comma-separated list of possible device names,
    e.g. sda,vda,hda rouchly matches the logic the Kilo ramdisk used.
  The resulting device WWN or serial (whatever is available) is then recorded
  in the root device hints (/properties/root_device) for a node.
--root-device-minimum-size minimum size of the considered devices
  The default value of 4 GiB matches what IPA does.
--overwrite-root-device-hints allows overwriting root device hints set
  previously. It's disabled by default to allow more precise control over the
  root device for some subset of nodes.

Note that for these arguments to work, this command should be run after
introspection. A separate documentation change will be posted for that.

Change-Id: I9f19554c5e7f34c8f63c1603c32b4d470fb12592