Completely removed the old discovery image support
This image is the private version of the deprecated bash-based introspection
ramdisk. It was never covered by any upstream CI. I doubt it even works now.
It was deprecated in Liberty and Mitaka. Time to remove it.

This change will allow to remove the unused (and probably broken) discovery
element from instack-undercloud.

Change-Id: I5668662e7ab5922adcda09c72c583e27e8b713b1