Add 'undercloud upgrade' command
The undercloud upgrade process entails updating the installed packages,
then running the installer again. However, after updating packages, a
number of services may no longer be able to run, as upon restart the
new code is not compatible with older settings. Therefore, this makes
the upgrade more 'atomic' by bundling the package update and the
undercloud install into a single command. An error (non-zero RC) in the
package update will halt the process.

Change-Id: I269ba919f1648f0e4b30ab710c386dafdad2a68d