Remove keystone init deprecation message
This shouldn't have been logged as deprecated until the new puppet
based init is merged.  Since at this point it doesn't look like
that will happen before Mitaka is branched, let's remove the
deprecation message so we don't cause confusion.

I'll push a revert follow-up that can be merged once this is
actually safe to deprecate.

Change-Id: I048f6403fe98d4384e4b4dd7033f59d1f6b636f3