Throw exception if undercloud password file is missing
utils.get_config_value() uses ConfigParser which treats non-existent
files as empty configuration. We are using this function to retrieve a
required value from the undercloud passwords file so failure to find the
file is actually a critical error. This patch adds a check and throws an
exception if the password file is missing.

Change-Id: I56cf8a99211cff7333e5f13539c3a145639686fa
Closes-Bug: #1536156