Use values from endpoint map for service endpoints
Instead of using values hardcoded in os-cloud-config. This commit
uses the values that come from the endpoint map that's used
in the t-h-t. This in turns, allows us to be able to change the
port and even the host used for the public and internal endpoints;
this is still not the case for the admin endpoints, as it is based
on the internal endpoint of the service. The aforementioned
limitation comes from os-cloud-config.

Change-Id: I0cf9ada1110f6b1452480039c136aee1fc3ccc82
Depends-On: I4cc5bbf2e7200f78cd90b93659c326a9200278d7
Depends-On: I07accff0914b93be09adf17a2552c5c922236c2e