Deprecate "baremetal show capabilities" command
This command was introduced to display node profiles. This job is now better
done by the "overcloud profiles list" command. As a result, this command is not
covered by our documentation at all.

Also, Ironic will have its own API and commands for handling capabilities
reasonably soon, so we risk clashing with them.

Change-Id: I423a03801280b9f9614214a7b8a615e8056b9b25