Allow import command to set deploy image and local boot
In our default flow we already have all the images uploaded by the time we
import the Ironic nodes. So we can make the "configure boot" command optional
by setting images and the local boot capability right during the import.

This change also introduces 2 more flags to opt-out from these changes:
--no-deploy-image will skip setting the deploy image
--instance-boot-option allows setting the boot_option capability explicitly
(it still requires changing flavors to match)

It's still possible to override the capabilities in the instackenv.json.

Change-Id: I00bbdff724ecfd6b840a74767617126734d7a32c