Raise instead of returning False when take_action fails
It looks like OSC is exiting with 0 if we return something from
take_action [1], no matter if we return True or False.

This submission changes the behaviour of the deploy command so
that it raises when something goes bad, instead of returning
False so that OSC will exit with non-0.

1. http://logs.openstack.org/45/289445/32/check-tripleo/gate-tripleo-ci-f22-upgrades/5bc5b3c/console.html#_2016-03-30_11_39_36_455

Change-Id: Ifa943f3f86eae5d81cbc900d4d6818619d73d3dc