Don't configure Nova v3 service
Microversions since Nova API v2.1 are aimed to replace the v3 work.

The /v2.1 is backwards compatible with the legacy /v2 endpoint.

What we called in the past /v3 is now something defunct in-tree. The /v2.1 API is based on the v3 work, but there are many things that differ, in particular with the backwards-compat thing I mentioned earlier.

We keep the /v2 path in api-paste.ini for making sure an upgrade doesn't trample operators and users but if you look in tree, that's redirecting to the v2.1 codepath (just not asking for microversions)

https://github.com/openstack/nova/blob/master/etc/nova/api-paste.ini#L18

In summary, we only need one endpoint, ie. /v2.1

Change-Id: Icdda1e08dd691aeca693725931709dcdc0efaf57
Related-Bug: #1564372