Generate ceph client key
By default only the admin user key is generated and this key is used
for both admin and openstack user.
Because the mode of the client's key file is 644, any user with a
valid shell on the controller/compute/ceph nodes can made admin
operations on the ceph cluster.
This patch allows to generate a random key for the openstack user that
will be used by tripleo-heat-templates.

Change-Id: I404665c09084f0a6cd2d8872940ee90220dc5f69
(cherry picked from commit adccde0cd572b5c4f2cc25d764afdcb9e1df5163)