Unset UpdateIdentifier on deploy/scale, pass in StackAction
Previously we tried to use UpdateIdentifier for two different things:
tell whether to perform package update, and also to tell whether the
top-level stack is being created or updated (which was incorrect and
resulted in bug 1567384, and an attempt to work around that bug resulted
in bug 1567385).

We cannot use Heat's "action" conditionals in some cases, because they
refer to the direct parent stack, which can yield undesirable results
when introducing new nested stacks or temporarily no-opping something
and then adding it back (in both these cases, "action" would be
considered "CREATE", even though the top-level stack is in "UPDATE").

So we clear UpdateIdentifier on deploy/scale to make sure that package
update doesn't get triggered, and we pass new StackAction parameter to
allow us to correctly distinguish between create vs. update of the
top-level stack, and condition the pacemaker restarts on this variable
instead.

Change-Id: I9dc3b4cd8a6a71df34d8babf0e4c6505041f5311
Related-Bug: #1567384
Closes-Bug: #1567385
(cherry picked from commit 4810728c3df4ebab3b8bd5a382ccdd4cd762ac55)