Make hypervisor stats check a validation
Moves the hypervisor stats check to be a proper validation. There is a
change in behavior with this move in that previously an exception was
raised and the deployment terminated when the check failed.

Moving it to a validation makes the check non-fatal by default, however
it can be made fatal again by specifying --validation-errors-fatal on
the cli.

Change-Id: I87eebcb8603de8620610cbbc8ca1f5c7d3aed992