Make node count a validation
Moves the node count check to be a proper validation. There is a change
in behavior with this move in that previously an exception was raised
and the deployment terminated when the check failed.

Moving it to a validation makes the check non-fatal by default, however
it can be made fatal again by specifying --validation-errors-fatal on
the cli.

Change-Id: Iffe51ad292d16b8273bf08e454d94adfe6ffbb33