Print success/failure message after image build
It's often not clear whether an image build failed or succeeded
just from looking at the final output from dib.  This change prints
an explicit success message when the images build successfully, and
raises an exception if we attempt the build but the image files
don't exist after. The latter is being done because there seem to be
cases where dib will return 0 but for one reason or another the
image files aren't actually created.  The explicit file check
should ensure that if the client reports all images built, then at
least the expected files are there.

Change-Id: I5e7372d79f75037873ffd9c0664c53ad2b44396b