Don't put --builder in image_group
image_group is a mutually exclusive argument group, which means
that if you pass one argument in it you can't pass any others. This
is clearly not what is intended for --builder since it needs to be
used with either --type or --all to specify the images to build.

Also adds a test case to verify the behavior.

Change-Id: I5db669fd1eff6720c009489f51b7ca2a3b2dcaca