Remove NeutronEnableTunnelling from python-tripleoclient
As described in bug 1532830, the OVS agent no longer uses the
enable_tunneling parameter which is controlled by
NeutronEnableTunneling, so NeutronEnableTunnelling will
be removed from Heat templates and puppet manifests. This
change removes the setting of NeutronEnableTunnelling
from the openstack command line in python-tripleoclient.
In conjunction with this, NeutronEnableTunneling will
be removed from tripleo-heat-templates.

Note that this change does not remove the command line
parameter 'neutron-disable-tunneling' in order to not
affect the UI and prevent script breakage. This parameter
is currently marked as [DEPRECATED] in the command line help.
A separate bug will be entered to either remove this parameter
or indicate that it will be ignored.

Change-Id: I28d33592374f60cb5222a866efaf9d137aca1c5a
Partial-Bug: 1532830