Unset UpdateIdentifier on deploy
We stopped clearing UpdateIdentifier because setting it to None caused
issues with template validation. However, this introduced a regression
with regards to minor updates.

We do patch stack-updates, which means that the UpdateIdentifer can
persist from previous runs. This can cause unwanted minor update on a
normal stack-update or on a scale-up, if the UpdateIdentifier is not
empty. We could prevent it on normal stack-update via some means,
e.g. persisting the last known update identifier value on the nodes and
comparing that with current value from the parameter. However, working
around this issue on scale-up would probably be considerably harder. A
good way forward seems to be to explicitly unset UpdateIdentifier as we
did before, but do it in a way that doesn't cause validation problems.

Change-Id: I7f6e3a65bb1f44a42cc532a537f1067bee168142
Closes-Bug: #1598124
Related-Bug: #1567385
Related-Bug: #1596640