Throw an error/exception when interactive overcloud update fails
The overcloud update currently returns an exit code of 0 regardless
of the outcome of the update. Unless the result is COMPLETE,
the client now throws an exception, and therefore a non-0 RC.

Also, change the unit tests to use 'COMPLETE' and not 'UPDATE_COMPLETE'
since Heat returns status as 'COMPLETE'.

Change-Id: Ib4aa7982718fda7c3b5fc4e5ea7501203c69bb98
Closes-Bug: 1599872