Make validation errors fatal by default
The current situation where errors and warnings are treated the same
does not make sense.  Errors should be things where the deployment
cannot continue.  If it's possible to deploy anyway, then that
should be a warning.

It could be argued that this should be notified for a cycle before
actually changing, but the current default is actually causing
deployers grief already because everyone has to remember to pass
--validation-errors-fatal to get sane behavior.  As noted in the
bug, the non-fatal default can have very bad consequences for users
who don't follow that best practice.  Essentially this is a CLI
bug that causes confusion and worse for users and a deprecation
cycle will only prolong that pain.  The only pain making the change
now should cause is possibly finding bugs in the validations, which
should be fixed anyway.

Change-Id: I9f77cd66547082b390c0c15034a2759e431045c5
Closes-Bug: 1608624