Remove excessive output when configuring nodes
The legacy 'baremetal configure boot' command calls the 'configure'
workflow individually for each node, causing a success/failure message
to be printed each time. The configured node UUID(s) are included in
the execution output, so printing a specific message for each node is
possible. However, considering that with many nodes this would make it
easy for failure messages to go unnoticed, and that the command prior
to the Mistral migration had no output, this patch removes the message
when nothing went wrong instead.

Change-Id: I3f1df8046ad92328fe0ee47fa073c58ca9ad5f20
Closes-Bug: #1619717