Generate Keystone credentials for overcloud
Keystone recently forced users to generate Keystone credentials in
/etc/keystone. The generation of credential can be done by
keystone-manage tool but it would require to do it on one node, export
the key into a storage system and collect the key on keystone servers.

This approach will generate a uniq Keystone credential key and export
two parameters: KeystoneCredential0 and KeystoneCredential1 that we will
use in puppet-keystone and Hiera to define the static content of the
keys.

It will allow us to do multinode deployments with the same keys.

This is a first iteration of Keystone credential supports. It doesn't
support rotations yet.

Change-Id: Ibca678b8bf222ecdef6615e34553598be3e55bf0
(cherry picked from commit 5f0694a64efcf6527306e8507efa92d03d60a05a)