Stop logging swiftclient errors
Swiftclient is logging a 404 when looking for
overcloud-without-mergepy.yaml.  To stop this, disable logging
below the critical level for swiftclient.  Note that we catch the
resulting exception from swiftclient and log it ourselves if
necessary, so there's no loss of information in the legitimate
error case.

Change-Id: I24144f3ab2b63c4a0f5fc2c9271e4a51a9ea1353
Closes-Bug: 1626267
(cherry picked from commit 4f9e9597041555c6d0657e9ccf544b781c0f4dcf)