Remove deprecation warning for bulk introspection
The new node management commands (bug 1595205) were meant to deprecate
some of the existing commands encroaching on the 'baremetal' namespace,
however there weren't all marked as such along the way nor were the
docs CI etc updated to match the new names and flow. Let's remove the
lonely deprecation warning to avoid the bad user experience of
following the current documentation yet getting warnings, and do the
deprecation properly at the beginning of Ocata.

This is not a backport because the warning is valid on master (Ocata)
and should only be removed from Newton.

Change-Id: I1cf4f4625d72330c87035f6ddcdef762d9930c34
Closes-Bug: #1626621